move: App.js内のMyTabsをAppNavigatorファイルに移動
